+++ lib-artnet/src/artnetparams.cpp
@@ -31,8 +31,10 @@
 # pragma GCC optimize ("Os")
 #endif
 
-#include <cstring>
-#include <cstdint>
+#include <stdint.h>
+#include <ctype.h>
+#include <string.h>
+#include <time.h>
 #include <cassert>
 
 #include "artnetparams.h"
@@ -63,7 +65,7 @@ ArtNetParams::ArtNetParams(ArtNetParamsStore *pArtNetParamsStore): m_pArtNetPara
 	m_tArtNetParams.nUniverse = 1;
 
 	for (uint32_t i = 0; i < ArtNet::MAX_PORTS; i++) {
-		m_tArtNetParams.nUniversePort[i] = static_cast<uint8_t>(1 + i);
+		m_tArtNetParams.nUniversePort[i] = 1 + i;
 	}
 
 	m_tArtNetParams.aOemValue[0] = ArtNetConst::OEM_ID[1];
@@ -159,8 +161,8 @@ void ArtNetParams::callbackFunction(const char *pLine) {
 	}
 
 	if (Sscan::HexUint16(pLine, ArtNetParamsConst::NODE_OEM_VALUE, nValue16) == Sscan::OK) {
-		m_tArtNetParams.aOemValue[0] = static_cast<uint8_t>(nValue16 >> 8);
-		m_tArtNetParams.aOemValue[1] = static_cast<uint8_t>(nValue16 & 0xFF);
+		m_tArtNetParams.aOemValue[0] = (nValue16 >> 8);
+		m_tArtNetParams.aOemValue[1] = (nValue16 & 0xFF);
 		m_tArtNetParams.nSetList |= ArtnetParamsMask::OEM_VALUE;
 		return;
 	}
@@ -238,7 +240,7 @@ void ArtNetParams::callbackFunction(const char *pLine) {
 				m_tArtNetParams.nUniversePort[i] = nValue8;
 				m_tArtNetParams.nSetList |= (ArtnetParamsMask::UNIVERSE_A << i);
 			} else {
-				m_tArtNetParams.nUniversePort[i] = static_cast<uint8_t>(i + 1);
+				m_tArtNetParams.nUniversePort[i] = i + 1;
 				m_tArtNetParams.nSetList &= ~(ArtnetParamsMask::UNIVERSE_A << i);
 			}
 			return;
@@ -272,9 +274,9 @@ void ArtNetParams::callbackFunction(const char *pLine) {
 			m_tArtNetParams.nDestinationIpPort[i] = nValue32;
 
 			if (nValue32 != 0) {
-				m_tArtNetParams.nMultiPortOptions |= static_cast<uint16_t>(ArtnetParamsMaskMultiPortOptions::DESTINATION_IP_A << i);
+				m_tArtNetParams.nMultiPortOptions |= (ArtnetParamsMaskMultiPortOptions::DESTINATION_IP_A << i);
 			} else {
-				m_tArtNetParams.nMultiPortOptions &= static_cast<uint16_t>(~(ArtnetParamsMaskMultiPortOptions::DESTINATION_IP_A << i));
+				m_tArtNetParams.nMultiPortOptions &= ~(ArtnetParamsMaskMultiPortOptions::DESTINATION_IP_A << i);
 			}
 		}
 	}
