+++ lib-dmxsend/include/h3/dmxsendmulti.h
@@ -2,7 +2,7 @@
  * @file dmxsendmulti.h
  *
  */
-/* Copyright (C) 2018-2021 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2018-2020 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -36,10 +36,10 @@ class DMXSendMulti: public DmxMulti, public LightSet  {
 public:
 	DMXSendMulti();
 
-	void Start(uint32_t nPort) override;
-	void Stop(uint32_t nPort) override;
+	void Start(uint8_t nPort) override;
+	void Stop(uint8_t nPort) override;
 
-	void SetData(uint32_t nPort, const uint8_t *pData, uint32_t nLength) override;
+	void SetData(uint8_t nPort, const uint8_t *pData, uint16_t nLength) override;
 
 	void Print() override;
